Completed the feature Continue build for Jenkins plugin. AB#2829 (#166)
Handled the exceptions in such a way that if scan time out occurs and continue build option is selected , and scan time out exception happens, succeeds the build. and last scan build report is shown.
For a new project , as previous reports are not available handled the scenario gracefully by logging the message for user